#1467 Try to clean up abstract message file handler hierarchy - Move some handling with the default file configuration down to the help message file handler since it is the only one with such a behavior now
